LayoutTests/fast/scrolling/mac/keyboard-scrolling-with-mouse-scroll.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true EventHandlerDrivenSmoothKeyboardScrollingEnabled=true ] -->

<html>

<head>
    <script src="../../../resources/ui-helper.js"></script>
    <script src="../../../resources/js-test.js"></script>
    <meta name="viewport" content="initial-scale=1.5, user-scalable=no">
    <script>
        if (window.testRunner) {
            testRunner.dumpAsText();
            testRunner.waitUntilDone();
        }

        let startPosition = 0;
        let delta = 0;
        let position = 0;
        let didScrollPastPosition = false;

        async function runTest()
        {
            eventSender.monitorWheelEvents();
            if (!window.testRunner || !testRunner.runUIScript)
                return;

            document.addEventListener('wheel', (event) => {
                startPosition = window.scrollY;
                // debug('startPosition = ' + startPosition);
                delta = event.deltaY;
                // debug('delta = ' + delta);

                document.addEventListener('scroll', (event) => {
                    position = window.scrollY;
                    // debug('position = ' + position);
                    if (startPosition + delta != position)
                        didScrollPastPosition = true;
                    // shouldBe('startPosition + delta', 'position');
                    // testRunner.notifyDone();
                })
            });

            await UIHelper.rawKeyDown("downArrow");
            
            setTimeout(async () => {
                await UIHelper.rawKeyUp("downArrow");
                await UIHelper.mouseWheelScrollAt(0, 0, 0, 0, 0, 1);

                await UIHelper.animationFrame();
                await UIHelper.animationFrame();
                await UIHelper.animationFrame();

                shouldBe('didScrollPastPosition', 'false')
                testRunner.notifyDone();
            }, 300);
        }
    </script>
</head>

<body onload="runTest()">
    <div style="height: 5000px;">
    </div>
</body>

</html>
